Remove upstream dependency on yarn
opensea-js fails to install unless yarn is present on the system due to this prepublish step being present in a git dependecy.  It would be more compatible to rely directly on the build script, or npm which always ships with node.

The best solution though, would be to publish a real package instead of consume via git.  

Thanks for the cool work!